Fix release Kotlin compile errors in rewards claim and FCM service
FieldValue.increment requires Long/Double, so the daily-reward coin credit
now passes a Long, and onNewToken uses applicationContext because `context`
is a Kotlin keyword and is not a property of FirebaseMessagingService.

+++ app/src/main/java/com/zerostress/manager/DailyLoginRewardsActivity.kt
@@ -233,7 +233,7 @@ private fun DailyLoginRewardsScreen() {
                             mapOf<String, Any>(
                                 "loginStreak" to newStreak,
                                 "lastLoginDate" to System.currentTimeMillis(),
-                                "coins" to com.google.firebase.firestore.FieldValue.increment(rewardCoins)
+                                "coins" to com.google.firebase.firestore.FieldValue.increment(rewardCoins.toLong())
                             )
                         )
                         true
